feat: viewmodel 분리
